+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- Add infrastructure packages
+- Add infrastructure packages (auth, http, event)
 - Add extensions package with support for FastAPI
 
 ### Improvements